src/requirements/base.txt
Django==1.10.2
psycopg2==2.6.2
python-social-auth==0.2.21
djrill==2.1.0
django-braces==1.10.0
